+++ plugins/postgres/dbt/include/postgres/macros/catalog.sql
@@ -36,7 +36,7 @@
         {%- endfor -%}
       )
       and not pg_is_other_temp_schema(sch.oid) -- not a temporary schema belonging to another session
-      and tbl.relpersistence = 'p' -- [p]ermanent table. Other values are [u]nlogged table, [t]emporary table
+      and tbl.relpersistence in ('p', 'u') -- [p]ermanent table or [u]nlogged table. Exclude [t]emporary tables
       and tbl.relkind in ('r', 'v', 'f', 'p') -- o[r]dinary table, [v]iew, [f]oreign table, [p]artitioned table. Other values are [i]ndex, [S]equence, [c]omposite type, [t]OAST table, [m]aterialized view
       and col.attnum > 0 -- negative numbers are used for system columns such as oid
       and not col.attisdropped -- column as not been dropped