+++ test/integration/074_postgres_unlogged_table_tests/models/table_unlogged.sql
@@ -1,0 +1,3 @@
+{{ config(materialized = 'table', unlogged = True) }}
+
+select 1 as column_a